Yeeee way to go ahead with chtrie
